Issue#SC-1388 added code logic to TenantMigrationActor
