+++ service/src/main/resources/application.conf
@@ -573,12 +573,6 @@ SunbirdMWSystem {
             nr-of-instances = 2
             dispatcher = rr-usr-dispatcher
           }
-        "/RequestRouter/*/MigrationActor"
-          {
-            router = smallest-mailbox-pool
-            nr-of-instances = 2
-            dispatcher = rr-usr-dispatcher
-          }
       }
     }
     remote {
